requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
whitenoise==6.0.0
django-rest-auth[with_social]
djangorestframework
django-rest-knox
Pillow
youtube-search-python
pafy
youtube_dl==2020.12.2
youtube-title-parse==1.0.0
websockets==10.0
django-webpack-loader==0.7.0
